Feat: 댓글 유효성 검사 기능 구현 완료 - 댓글 작성시 매칭글 ID, 댓글 내용 필수 입력 유효성 검사 - 댓글 수정 및 삭제시 자신의 댓글인지 확인하는 유효성 검사 기능 구현

+++ src/main/java/com/likelion/oegaein/domain/matching/service/CommentService.java
@@ -7,8 +7,10 @@ import com.likelion.oegaein.domain.matching.repository.CommentRepository;
 import com.likelion.oegaein.domain.matching.repository.MatchingPostRepository;
 import com.likelion.oegaein.domain.member.entity.profile.Member;
 import com.likelion.oegaein.domain.member.repository.MemberRepository;
+import com.likelion.oegaein.domain.member.validation.MemberValidator;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,16 +18,24 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class CommentService {
+    // constants
+    private final String NOT_FOUND_MATCHING_POST_ERR_MSG = "찾을 수 없는 매칭글입니다.";
+    private final String NOT_FOUND_MEMBER_ERR_MSG = "찾을 수 없는 사용자입니다.";
+    private final String NOT_FOUND_COMMENT_ERR_MSG = "찾을 수 없는 댓글입니다.";
+    // repository
     private final CommentRepository commentRepository;
     private final MatchingPostRepository matchingPostRepository;
     private final MemberRepository memberRepository;
+    // validator
+    private final MemberValidator memberValidator;
 
     @Transactional
-    public CreateCommentResponse saveComment(CreateCommentData dto){
+    public CreateCommentResponse saveComment(CreateCommentData dto, Authentication authentication){
         // find author, parentComment, matchingPost and receiver
         MatchingPost matchingPost = matchingPostRepository.findById(dto.getMatchingPostId())
-                .orElseThrow(() -> new EntityNotFoundException("Not Found: matchingPost"));
-        Member author = memberRepository.findById(2L).orElseThrow(() -> new EntityNotFoundException("Not Found: member")); // 임시 작성자
+                .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_MATCHING_POST_ERR_MSG));
+        Member author = memberRepository.findByEmail(authentication.getName())
+                .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_MEMBER_ERR_MSG)); // 임시 작성자
 
         // create new comment
         Comment comment = Comment.builder()
@@ -39,17 +49,23 @@ public class CommentService {
     }
 
     @Transactional
-    public UpdateCommentResponse updateComment(UpdateCommentData dto, Long commentId){
+    public UpdateCommentResponse updateComment(UpdateCommentData dto, Long commentId, Authentication authentication){
+        Member authenticatedMember = memberRepository.findByEmail(authentication.getName())
+                .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_MEMBER_ERR_MSG));
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new EntityNotFoundException("Not Found: comment"));
+                .orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_COMMENT_ERR_MSG));
+        memberValidator.validateIsOwnerComment(authenticatedMember.getId(), comment.getAuthor().getId());
         comment.updateContent(dto.getContent());
         return new UpdateCommentResponse(commentId);
     }
 
     @Transactional
-    public DeleteCommentResponse removeComment(Long commentId){
+    public DeleteCommentResponse removeComment(Long commentId, Authentication authentication){
+        Member authenticatedMember = memberRepository.findByEmail(authentication.getName())
+                .orElseThrow(() -> new EntityNotFoundException(NOT_FOUND_MEMBER_ERR_MSG));
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new EntityNotFoundException("Not Found: comment"));
+                .orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_COMMENT_ERR_MSG));
+        memberValidator.validateIsOwnerComment(authenticatedMember.getId(), comment.getAuthor().getId());
         if(comment.getReplies().isEmpty()){ // no child comments
             commentRepository.delete(comment);
         }else{ // has child comments
